requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
email-validator==1.3.1
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
# Removed psycopg2-binary as it requires PostgreSQL installation
# For production, uncomment: psycopg2-binary==2.9.7
# Removed tensorflow as it's a large dependency and not needed for initial setup
# For AI training, uncomment: tensorflow==2.14.0
numpy==1.24.3
pandas==2.1.0
